+++ src/lib/st-tooltip/st-tooltip.component.spec.ts
@@ -8,92 +8,77 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
+import { Component, DebugElement, Input } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Component } from '@angular/core';
+import { By } from '@angular/platform-browser';
 
 import { StTooltipComponent } from './st-tooltip.component';
-
-@Component({
-   template: `<span st-tooltip></st-tooltip>`
-})
-class TestStTooltipComponent { }
 
 const id: string = 'tooltipId';
 const originalContent: string = 'This is the original element content';
 const tooltipText: string = 'This text will be displayed in our tooltip';
 
-let component: StTooltipComponent;
-let fixture: ComponentFixture<TestStTooltipComponent>;
-let nativeElement: any;
-let template: string = '';
-
-function createTestComponent(customTemplate?: string): Promise<ComponentFixture<TestStTooltipComponent>> {
-   if (customTemplate) {
-      TestBed.overrideComponent(TestStTooltipComponent, {
-         set: {
-            template: customTemplate
-         }
-      });
-   }
-   return TestBed.compileComponents();
+@Component({
+   template: '<span id="' + id + '" st-tooltip [attr.title]="tooltipText" [showOnClick]="showOnClick">' + originalContent + '</span>'
+})
+class TestStTooltipComponent {
+   @Input() showOnClick: boolean = false;
+   @Input() tooltipText: string = tooltipText;
 }
+
+let component: TestStTooltipComponent;
+let element: DebugElement;
+let fixture: ComponentFixture<TestStTooltipComponent>;
 
 describe('StTooltip', () => {
 
    beforeEach(async(() => {
-      TestBed.configureTestingModule({
+      TestBed
+      .configureTestingModule({
          declarations: [
             StTooltipComponent,
             TestStTooltipComponent
          ]
-      });
+      })
+      .compileComponents();
    }));
 
-   it('It has to display content inside a span without a title', async(() => {
-      template = '<span st-tooltip id="' + id + '" title="' + tooltipText + '">' + originalContent + '</span>';
-      createTestComponent(template).then(() => {
-         fixture = TestBed.createComponent(TestStTooltipComponent);
-         nativeElement = fixture.nativeElement;
-         fixture.detectChanges();
+   beforeEach(() => {
+      fixture = TestBed.createComponent(TestStTooltipComponent);
+      component = fixture.componentInstance;
+      element = fixture.debugElement.query(By.css('#' + id));
+      fixture.detectChanges();
+   });
 
-         let tooltip: Element = nativeElement.querySelector('.st-tooltip');
-         expect(tooltip).toBeDefined();
 
-         let span: Element = tooltip.querySelector('span');
-         expect(span.getAttribute('title')).toBe('');
-         expect(span.innerHTML).toContain(originalContent);
-      });
-   }));
+   it('It has to display content inside a span without a title', () => {
+      let span: Element = element.query(By.css('span')).nativeElement;
+      expect(span.getAttribute('title')).toBe('');
+      expect(span.innerHTML).toContain(originalContent);
+   });
 
-   it('It can be configured to be displayed and hidden when user clicks on its associated content', async(() => {
-      template = '<span st-tooltip id="' + id + '" title="' + tooltipText + '">' + originalContent + '</span>';
-      createTestComponent(template).then(() => {
-         fixture = TestBed.createComponent(TestStTooltipComponent);
-         nativeElement = fixture.nativeElement;
-         fixture.detectChanges();
+   it('It has st-tooltip class only if host element has a title', () => {
+      expect(element.nativeElement.title).toBe(tooltipText);
+      expect(element.nativeElement.className).toBe('st-tooltip');
 
-         let tooltip: Element = nativeElement.querySelector('.st-tooltip');
-         expect(tooltip).toBeDefined();
+      component.tooltipText = '';
+      fixture.detectChanges();
+      expect(element.nativeElement.title).toBe('');
+      expect(element.nativeElement.className).toBe('');
+   });
 
-         let span: Element = tooltip.querySelector('span');
-         expect(span.getAttribute('title')).toBe('');
-         expect(span.innerHTML).toContain(originalContent);
-      });
-   }));
+   it('If showOnClick, st-tooltip-on class is added or removed on click', () => {
+      component.showOnClick = true;
+      fixture.detectChanges();
+      expect(element.nativeElement.className).toBe('');
 
-   it('It can be configured to be displayed and hidden when user puts the mouse on its associated content', async(() => {
-      template = '<span st-tooltip id="' + id + '" title="' + tooltipText + '">' + originalContent + '</span>';
-      createTestComponent(template).then(() => {
-         fixture = TestBed.createComponent(TestStTooltipComponent);
-         nativeElement = fixture.nativeElement;
-         fixture.detectChanges();
+      element.nativeElement.click();
+      fixture.detectChanges();
+      expect(element.nativeElement.className).toBe('st-tooltip-on');
 
-         let tooltip: Element = nativeElement.querySelector('.st-tooltip');
-         expect(tooltip).toBeDefined();
+      fixture.nativeElement.click();
+      fixture.detectChanges();
+      expect(element.nativeElement.className).toBe('');
+   });
 
-         let span: Element = tooltip.querySelector('span');
-         expect(span.getAttribute('title')).toBe('');
-         expect(span.innerHTML).toContain(originalContent);
-      });
-   }));
 });